Refactoring to use long and bit counts for minutes. Bumping version.

+++ Brizbee.Api/Serialization/Expanded/PunchExpanded.cs
@@ -92,9 +92,9 @@ namespace Brizbee.Api.Serialization.Expanded
 
         public int? Punch_PayrollRateId { get; set; }
 
-        public int Punch_Minutes { get; set; }
+        public long Punch_Minutes { get; set; }
 
-        public int Punch_CumulativeMinutes { get; set; }
+        public long Punch_CumulativeMinutes { get; set; }
 
 
         // Customer Details
